Added check for `data` input type
An initial step has been added to check that the input `data` is either a `lk.TargetPixelFileCollection` with all items being `lk.TessTargetPixelFile` objects, or the `data` is a `lk.LightCurveCollection` with all items being `lk.TessLightCurve` objects. All type checks have also been consolidated to the setup section of the code.